Fixes offending flake8/mastool errors
HEAD is not passing the tox job. This commit fixes 4 violations. Note
that I removed the generic try/catch exception as it is a very bad
practice. You should always name the exception you're catching and avoid
catching generic ones (cf M005 in mastool).

If you don't know which exceptions might be raised, it's generally
better to let the exception propagate and send the Traceback to the user.