Improves error handling
In Python you really want to avoid using return values to indicate
success or failure. Instead it's easier to just go ahead and ignore
errors, while catching possible exception from the calling
function. `return Boolean` should be reserved for function which only
compute a predicate.

This patch will review error handling by introducing better usage of
exceptions. It also improves a bit on the unit tests to take the change
into account. (`assertRaises` is more expressive than `assertFalse`).

Finally I've also started removing comments I found to be useless and
not adding much to the understanding of the script. In later patches I
suggest we do a full file sweep to remove comments as much as we can.